+++ src/dos/dos_programs.cpp
@@ -601,7 +601,7 @@ void MenuBrowseImageFile(char drive, bool arc, bool boot, bool multiple) {
         if (lTheOpenFileName) fname = GetNewStr(lTheOpenFileName);
         if (multiple&&fname.size()) {
             files += "\"";
-            for (int i=0; i<fname.size(); i++)
+            for (size_t i=0; i<fname.size(); i++)
                 files += fname[i]=='|'?"\" \"":std::string(1,fname[i]);
             files += "\" ";
         }
@@ -610,7 +610,7 @@ void MenuBrowseImageFile(char drive, bool arc, bool boot, bool multiple) {
             if (lTheOpenFileName) {
                 fname = GetNewStr(lTheOpenFileName);
                 files += "\"";
-                for (int i=0; i<fname.size(); i++)
+                for (size_t i=0; i<fname.size(); i++)
                     files += fname[i]=='|'?"\" \"":std::string(1,fname[i]);
                 files += "\" ";
             }
